Tests/LibWeb/Text/expected/css/inline-relative-css-image-base-url.txt

Computed URL before base change: url("../wpt-import/images/anim-gr.gif")
Computed URL after base change: url("../wpt-import/images/anim-gr.gif")
active animations before base change: 1
active animations after base change: 1
